Debug mode added

+++ src/implicit_neural_representation.py
@@ -1,3 +1,4 @@
+import argparse
 import random
 from datetime import datetime
 
@@ -7,16 +8,23 @@ from torch.utils.tensorboard import SummaryWriter
 from inr_model import INR
 from utils import *
 
-UNIFORM_TRAINING_EPOCHS = 10
+parser = argparse.ArgumentParser()
+parser.add_argument("-d", "--debug", dest="debug", help="Enable debug mode", action="store_true", default=False)
+args = parser.parse_args()
+
+UNIFORM_TRAINING_EPOCHS = 15
 GRADIENT_BASED_TRAINING_EPOCHS = 5
 INTRA_RAY_DEGREES = 1
+
+debug = args.debug
+
+if debug:
+    print("---DEBUG MODE ACTIVATED---")
 
 print(f"Started {datetime.now().strftime('%Y/%m/%d %H:%M:%S')}")
 
 torch.manual_seed(41)
-
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
 torch.set_default_device(device)
 model = INR(device=device)
 model.to(device)
@@ -28,12 +36,17 @@ def train_one_epoch(epoch_index, tb_writer):
     running_loss = 0.
     last_loss = 0.
 
-    for val_index in range(100):  # batches
-        laser_rays = np.zeros((500, 500, 1), np.uint8)
-        laser_points = np.zeros((500, 500, 1), np.uint8)
-        laser_points.fill(255)
-        laser_points_after_knn = np.zeros((500, 500, 1), np.uint8)
-        laser_points_after_knn.fill(255)
+    for val_index in range(50):  # batches
+        if debug:
+            laser_rays = np.zeros((500, 500, 1), np.uint8)
+            laser_points = np.zeros((500, 500, 1), np.uint8)
+            laser_points.fill(255)
+            laser_points_after_knn = np.zeros((500, 500, 1), np.uint8)
+            laser_points_after_knn.fill(255)
+        else:
+            laser_rays = None
+            laser_points = None
+            laser_points_after_knn = None
 
         external = []
         internal = []
@@ -41,15 +54,16 @@ def train_one_epoch(epoch_index, tb_writer):
 
         # angles = random.sample(range(0, 360), 300)
         inputs = np.array([]).reshape(0, 2)
-        for _ in range(300):
+        for _ in range(400):
             start_point = [250, 250]  # [random.randint(0, 500), random.randint(0, 500)]
             angle = random.uniform(0., 360.)
 
             # direction = 1 if random.random() >= 0.5 else -1
-            simulate_laser_ray(start_point, angle, 1, laser_rays)
+            if debug:
+                simulate_laser_ray(start_point, angle, 1, laser_rays)
             e, inner, u = generate_laser_points(start_point, angle)
             external.extend(random.sample(e, 40 if len(e) > 40 else len(e)))
-            internal.extend(random.sample(inner, 40 if len(inner) > 40 else len(inner)))
+            internal.extend(random.sample(inner, 10 if len(inner) > 10 else len(inner)))
             unknown.extend(random.sample(u, 40 if len(u) > 40 else len(u)))
 
             # inputs = np.concatenate((inputs, random.sample(e, 64)), axis=0)
@@ -57,18 +71,19 @@ def train_one_epoch(epoch_index, tb_writer):
 
         external, internal = knn_point_classification(external, internal, unknown, 5)
 
-        for e in list(filter(lambda p: 0 < p[0] < 500 and 0 < p[1] < 500, external)):
-            laser_points_after_knn[e[1], e[0]] = 150
-            # laser_points[e[1], e[0]] = 0
+        if debug:
+            for e in list(filter(lambda p: 0 < p[0] < 500 and 0 < p[1] < 500, external)):
+                laser_points_after_knn[e[1], e[0]] = 150
+                # laser_points[e[1], e[0]] = 0
 
-        for inne in list(filter(lambda p: 0 < p[0] < 500 and 0 < p[1] < 500, internal)):
-            laser_points[inne[1], inne[0]] = 0
-            laser_points_after_knn[inne[1], inne[0]] = 0
+            for inne in list(filter(lambda p: 0 < p[0] < 500 and 0 < p[1] < 500, internal)):
+                laser_points[inne[1], inne[0]] = 0
+                laser_points_after_knn[inne[1], inne[0]] = 0
 
-        cv.imshow("laser negative points", laser_points)
-        cv.imshow("laser negative points after knn", laser_points_after_knn)
-        cv.imshow("laser rays", laser_rays)
-        cv.waitKey(1)
+            cv.imshow("laser negative points", laser_points)
+            cv.imshow("laser negative points after knn", laser_points_after_knn)
+            cv.imshow("laser rays", laser_rays)
+            cv.waitKey(1)
 
         inputs = np.concatenate((inputs, external), axis=0)
         inputs = np.concatenate((inputs, internal), axis=0)
@@ -89,6 +104,8 @@ def train_one_epoch(epoch_index, tb_writer):
         optimizer.step()
 
         running_loss += loss.item()
+
+        print(f"batch {val_index} done", end='\r')
         if val_index % 1000 == 999:
             last_loss = running_loss / 1000  # loss per batch
             print('  batch {} loss: {}'.format(val_index + 1, last_loss))
@@ -103,14 +120,18 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
     running_loss = 0.
     last_loss = 0.
 
-    for val_index in range(50):  # batches
-        laser_points = np.zeros((500, 500, 1), np.uint8)
+    for val_index in range(10):  # batches
+        if debug:
+            laser_points = np.zeros((500, 500, 1), np.uint8)
+        else:
+            laser_points = None
+
         external = []
         internal = []
         unknown = []
 
         inputs = np.array([]).reshape(0, 2)
-        for _ in range(200):
+        for _ in range(100):
             start_point = [250, 250]
             angle = random.uniform(0., 360.)
 
@@ -130,7 +151,7 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
             distribution = gradient_values.flatten()
 
             for sampled in np.random.choice(np.array([d for d in range(len(distribution))]),
-                                            size=(20 if len(t) > 20 else len(t)),
+                                            size=(10 if len(t) > 10 else len(t)),
                                             p=distribution, replace=False):
                 external.append(e[sampled])
 
@@ -148,7 +169,7 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
             distribution = gradient_values.flatten()
 
             for sampled in np.random.choice(np.array([d for d in range(len(distribution))]),
-                                            size=(20 if len(t) > 20 else len(t)),
+                                            size=(10 if len(t) > 10 else len(t)),
                                             p=distribution, replace=False):
                 internal.append(inner[sampled])
             ###
@@ -165,7 +186,7 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
             distribution = gradient_values.flatten()
 
             for sampled in np.random.choice(np.array([d for d in range(len(distribution))]),
-                                            size=(20 if len(t) > 20 else len(t)),
+                                            size=(10 if len(t) > 10 else len(t)),
                                             p=distribution, replace=False):
                 unknown.append(u[sampled])
 
@@ -176,15 +197,16 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
         inputs = np.concatenate((inputs, external), axis=0)
         inputs = np.concatenate((inputs, internal), axis=0)
 
-        for p in inputs:
-            px = round(p[1])
-            py = round(p[0])
-            if 0 <= px < 500 and 0 <= py < 500:
-                laser_points[px, py] = 255
+        if debug:
+            for p in inputs:
+                px = round(p[1])
+                py = round(p[0])
+                if 0 <= px < 500 and 0 <= py < 500:
+                    laser_points[px, py] = 255
 
-        cv.imshow('gradient based sampling', laser_points)
-        cv.imwrite(f'extracted_{epoch_index}.png', laser_points)
-        cv.waitKey(1)
+            cv.imshow('gradient based sampling', laser_points)
+            cv.imwrite(f'extracted_{epoch_index}.png', laser_points)
+            cv.waitKey(1)
 
         labels = torch.tensor([[1] for _ in external] + [[-1] for _ in internal], dtype=torch.float32,
                               requires_grad=True, device=device)
@@ -195,6 +217,7 @@ def train_one_gradient_based_epoch(epoch_index, tb_writer):
         loss.backward()
         optimizer.step()
 
+        print(f"batch {val_index} done", end='\r')
         running_loss += loss.item()
         if val_index % 1000 == 999:
             last_loss = running_loss / 1000  # loss per batch
@@ -278,10 +301,11 @@ for epoch in range(GRADIENT_BASED_TRAINING_EPOCHS):
     gradient_image /= gradient_sum
     flattened_distribution = gradient_image.flatten()
 
-    image = gradient_image * 255.
-    cv.imshow("gradient", image)
-    cv.imwrite(f"images/gradient_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}-{epoch_number}.png", image)
-    cv.waitKey(1)
+    if debug:
+        image = gradient_image * 255.
+        cv.imshow("gradient", image)
+        cv.imwrite(f"images/gradient_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}-{epoch_number}.png", image)
+        cv.waitKey(1)
 
     # Make sure gradient tracking is on, and do a pass over the data
     model.train(True)
@@ -327,9 +351,9 @@ for epoch in range(GRADIENT_BASED_TRAINING_EPOCHS):
     epoch_number += 1
 
 print(f"done {datetime.now().strftime('%Y/%m/%d %H:%M:%S')}")
-# cv.waitKey(0)
-cv.destroyAllWindows()
-
 model.train(False)
-
 torch.save(model.state_dict(), 'model')
+
+if debug:
+    cv.waitKey(0)
+    cv.destroyAllWindows()
